# Improve performance of value(NonlinearExpression)