@odow 2021-03-02
With the DFS, the order_subexpressions is fixed:
```julia
julia> @time main(500);
  1.497348 seconds (2.21 M allocations: 977.662 MiB, 41.23% gc time)
```