@odow 2021-03-03
And now
```julia
julia> @time main(500);
  1.071244 seconds (1.80 M allocations: 930.415 MiB, 36.37% gc time)
```