@odow 2021-03-07
Following @mlubin's suggestions, we now have:
```
  1.291303 seconds (1.87 M allocations: 1.301 GiB, 34.93% gc time)
```
A bit slower than before, but not by much, and it should scale much more gracefully.