@odow 2021-03-07
Now with much updated docstrings.